Share some inspector timeline components
