+++ apps/webapp/app/components/runs/v3/InspectorTimeline.tsx
@@ -1,0 +1,63 @@
+import { ReactNode } from "react";
+import { cn } from "~/utils/cn";
+
+type RunTimelineItemProps = {
+  title: ReactNode;
+  subtitle?: ReactNode;
+  state: "complete" | "error";
+};
+
+export function RunTimelineEvent({ title, subtitle, state }: RunTimelineItemProps) {
+  return (
+    <div className="grid h-5 grid-cols-[1.125rem_1fr] text-sm">
+      <div className="flex items-center justify-center">
+        <div
+          className={cn(
+            "size-[0.3125rem] rounded-full",
+            state === "complete" ? "bg-success" : "bg-error"
+          )}
+        ></div>
+      </div>
+      <div className="flex items-baseline justify-between gap-3">
+        <span className="font-medium text-text-bright">{title}</span>
+        {subtitle ? <span className="text-xs text-text-dimmed">{subtitle}</span> : null}
+      </div>
+    </div>
+  );
+}
+
+type RunTimelineLineProps = {
+  title: ReactNode;
+  state: "complete" | "delayed" | "inprogress";
+};
+
+export function RunTimelineLine({ title, state }: RunTimelineLineProps) {
+  return (
+    <div className="grid h-6 grid-cols-[1.125rem_1fr] text-xs">
+      <div className="flex items-stretch justify-center">
+        <div
+          className={cn(
+            "w-px",
+            state === "complete" ? "bg-success" : state === "delayed" ? "bg-text-dimmed" : ""
+          )}
+          style={
+            state === "inprogress"
+              ? {
+                  width: "1px",
+                  height: "100%",
+                  background:
+                    "repeating-linear-gradient(to bottom, #3B82F6 0%, #3B82F6 50%, transparent 50%, transparent 100%)",
+                  backgroundSize: "1px 6px",
+                  maskImage: "linear-gradient(to bottom, black 50%, transparent 100%)",
+                  WebkitMaskImage: "linear-gradient(to bottom, black 50%, transparent 100%)",
+                }
+              : undefined
+          }
+        ></div>
+      </div>
+      <div className="flex items-center justify-between gap-3">
+        <span className="text-text-dimmed">{title}</span>
+      </div>
+    </div>
+  );
+}
